--- bug-tracker-backend/handler.go
@@ -1,1 +1,0 @@
-
