+++ bug-tracker-backend/src/conf/database.go
