+++ bug-tracker-backend/src/controllers/bug_controller.go
@@ -9,10 +9,10 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/lib/pq"
 
-	"github.com/WNBARookie/BugTracker/bug-tracker-backend/internal/conf"
-	"github.com/WNBARookie/BugTracker/bug-tracker-backend/internal/models"
-	types "github.com/WNBARookie/BugTracker/bug-tracker-backend/internal/types"
-	"github.com/WNBARookie/BugTracker/bug-tracker-backend/internal/utils"
+	"github.com/WNBARookie/BugTracker/bug-tracker-backend/src/conf"
+	"github.com/WNBARookie/BugTracker/bug-tracker-backend/src/models"
+	types "github.com/WNBARookie/BugTracker/bug-tracker-backend/src/types"
+	"github.com/WNBARookie/BugTracker/bug-tracker-backend/src/utils"
 )
 
 func CreateBug(c *gin.Context) {
