+++ bug-tracker-backend/src/middlewares/auth_middleware.go
@@ -10,7 +10,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/golang-jwt/jwt/v5"
 
-	"github.com/WNBARookie/BugTracker/bug-tracker-backend/internal/utils"
+	"github.com/WNBARookie/BugTracker/bug-tracker-backend/src/utils"
 )
 
 func RequireAuth(c *gin.Context) {
