+++ bug-tracker-backend/src/types/user.go
