+++ bug-tracker-backend/src/utils/context_extractors.go
@@ -3,7 +3,7 @@ package utils
 import (
 	"github.com/gin-gonic/gin"
 
-	"github.com/WNBARookie/BugTracker/bug-tracker-backend/internal/models"
+	"github.com/WNBARookie/BugTracker/bug-tracker-backend/src/models"
 )
 
 func ExtractUserFromContext(c *gin.Context) models.User {
